// app/(manager)/screens/types.ts
  export interface EmployerDetailsDisplayProps {
    id: string;
        name: string;
        email: string;
        phone: string;
        homeAddress: string;
        dateOfBirth: string;
        Language: string;
        identityVerificationStatus: string;
        backgroundCheckStatus: string;
        description: string;
        rating: number;
        assurance: number;
        urlToImage: any;
        imageStatus: string;
        IDCardStatus: string;
        PassportStatus: string;
        otherCertificateStatus: string;
  }
// ...
export interface ShiftDetailsProps {
  id: string,
  shiftType: string;
  totalEarnings: string;
  date: string;
  time: string;
  endTime: string;
  hourlyRate: string;
  numberOfHours: string;
  numberOfOpenings: string;
  location: string;
  description: string;
  employee: string;
  employer: string;
  employerId: string;
  employeeId: string;
  backgroundImage: string;
}

export interface UserDataProps {
      id: string;
      name: string;
      email: string;
      phone: string;
      homeAddress: string;
      dateOfBirth: string;
      Language: string;
      identityVerificationStatus: string;
      backgroundCheckStatus: string;
      description: string;
      rating: string,
      assurance: string,
      urlToImage: any,
      imageStatus: string;
      IDCardStatus: string;
      PassportStatus: string;
      otherCertificateStatus: string;
  }